error handle

+++ src/main/java/study/community/controller/PublishController.java
@@ -43,7 +43,7 @@ public class PublishController {
     public String create(@RequestParam(value = "title", required = false) String title,
                          @RequestParam(value = "description", required = false)String descrition,
                          @RequestParam(value = "tag", required = false) String tag,
-                         @RequestParam(value = "id",required = false) Integer id,
+                         @RequestParam(value = "id",required = false) Long id,
                          HttpServletRequest request,
                          Model model) {
         model.addAttribute("title", title);
@@ -81,7 +81,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id")Integer id,Model model) {
+    public String edit(@PathVariable(name="id")Long id,Model model) {
         QuestionDto quesiton = questionService.getById(id);
         model.addAttribute("title",quesiton.getTitle());
         model.addAttribute("description", quesiton.getDescription());
